qw: align video, sound and input code with qk1
note that sound will cut up by default unless devaudio delay is increased, or
cl_maxfps or rate are adjusted for a maximum framerate of 72 (hardcoded).
the latter is recommended.  no idea why one would want a lower framerate.

video code also essentially assumes that the screen is always updated in its
entirety.  even if this is not always true, it matters only ingame, where it
(currently) is.