increase default memsize to 64mb
it is unlikely that qk1 would actually be run on a machine
with constrained resources at this point.  in fact, people
often suggest a memsize of 192, but it's unclear whether
these mods *actually* need that much memory.  in any case
this avoids stupid errors people could have at first
contact.