snd: fix playback from physical cd's and put it in its own proc
this is how playback from cd's is supposed to work, reading
from the cd without waiting for the rest of the code and mixing
it with the sound effects, and it also drastically improves
performance.  the rest of the program shouldn't have to wait
for the cd drive, and it's fine to leave mixing to mixfs rather
than having to sync and mix things ourselves (ioproc or
otherwise).