vid (sdl): fix weird glitches and crashes
Hello, Wayland.

Seems like doing things in the previous order and not "grabbing" the
mouse (setting it to relative mode) resulted in somehow view getting
stuck randomly (moving the camera in one direction would pull it back
in a fraction of a second) and crashes (in random unrelated places,
and it did not crash with valgrind). Just pressing the right mouse
button once, before *moving* the mouse at all, would result in glitches.

Solve this by changing the order, keeping sway from resizing the window
on start, and grabbing the mouse right away.