fix the usual SZ_GetSpace overflow on huge maps
This seems to fix maps like 'telefragged' (did not even spawn before)
and 'ad_tears' (would break in the middle of the game).

Keep an array of signons and move data to the newly allocated sizebuf
if it doesn't fully fit in the old one. A bit better(?) implementation
compared to other engines as you don't have to figure out the size of
the frame by yourself.